Fix local development on macOS
Since macOS Monterey in 2001, the Control Center service has been
using port 5000 at startup[1], meaning that the moto test server will
be binding to an already-in-use port and silently fail, causing
most tests to fail with a somewhat inexplicable 403 Forbidden.

This fixes that by making the port a fixture in tests/fixtures and
using that constant anywhere the test code was using port 5000.